requirements.txt
flask
requests
flask-cors
cairosvg
pillow
openai
python-dotenv
gunicorn
opencv-python-headless
numpy
pytesseract
vtracer==0.6.11